added controllers

+++ HRMS.Domain/Services/ContractService.cs
@@ -18,6 +18,7 @@ namespace HRMS.Domain.Services
             _mapper = mapper;
         }
 
+
         public async Task<IEnumerable<ContractDTO>> GetContracts()
         {
             var contractsEntity = await _contractRepository.GetContracts();
@@ -30,24 +31,31 @@ namespace HRMS.Domain.Services
             return _mapper.Map<ContractDTO>(contractEntity);
         }
 
-        public async Task Add(ContractDTO contract)
+        public async Task Add(ContractDTO contractDTO)
         {
             // converter o DTO antes - lógica inversa
-            var contractEntity = _mapper.Map<Contract>(contract);
+            var contractEntity = _mapper.Map<Contract>(contractDTO);
             await _contractRepository.Create(contractEntity);
         }
 
-        public async Task Update(ContractDTO contract)
+        public async Task<bool> Update(ContractDTO contractDTO)
         {
-            var contractEntity = _mapper.Map<Contract>(contract);
+            var contractEntity = _mapper.Map<Contract>(contractDTO);
+            if (contractEntity == null)
+                throw new Exception("Contract not found");
+
             await _contractRepository.Update(contractEntity);
+            return true;
         }
 
-        public async Task Remove(int id)
+        public async Task<bool> Remove(int id)
         {
-            // preciso localizar o objeto antes - apenas task não retorna valor
-            var contractEntity = _contractRepository.GetById(id).Result;
+            var contractEntity = await _contractRepository.GetById(id);
+            if (contractEntity == null)
+                throw new Exception("Contract not found");
             await _contractRepository.Remove(contractEntity);
+            return true;
+            
         }
     }
 }
