+++ HRMS.Domain/Services/EmployeeService.cs
@@ -35,23 +35,32 @@ namespace HRMS.Domain.Services
             return _mapper.Map<EmployeeDTO>(employeeEntity);
         }
 
-        public async Task Add(EmployeeDTO employee)
+        public async Task Add(EmployeeDTO employeeDTO)
         {
-           var employeeEntity = _mapper.Map<Employee>(employee);
+           var employeeEntity = _mapper.Map<Employee>(employeeDTO);
             await _employeeRepository.CreateAsync(employeeEntity);
         }
 
-        public async Task Update(EmployeeDTO employee)
+        public async Task<bool> Update(EmployeeDTO employeeDTO)
         {
-            var employeeEntity = _mapper.Map<Employee>(employee);
-            await _employeeRepository.UpdateAsync(employeeEntity);
+            var employeeEntity = _mapper.Map<Employee>(employeeDTO);
+            var employee = await _employeeRepository.UpdateAsync(employeeEntity);
+            if (employeeEntity == null)
+                throw new Exception("Employee not found");
+            return true;
         }
 
 
-        public async Task Remove(int id)
+        public async Task<bool> Remove(int id)
         {
-            var employeeEntity = _employeeRepository.GetByIdAsync(id).Result;
+            //var employeeEntity = _employeeRepository.GetByIdAsync(id).Result;
+            //await _employeeRepository.RemoveAsync(employeeEntity);
+
+            var employeeEntity = await _employeeRepository.GetByIdAsync(id);
+            if (employeeEntity == null)
+                throw new Exception("Employee not found");
             await _employeeRepository.RemoveAsync(employeeEntity);
+            return true;
         }
 
     }
